LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-scroll-shy-target-expected.txt

This tests gesture scrolling div which is deleted while latched, Square is (mostly) green on pass

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
wheel event 0+> [object HTMLDivElement]
wheel event 1+> [object HTMLDivElement]
wheel event 2+> [object HTMLDivElement]
PASS movingdiv.scrollTop is 62
PASS movingdiv.scrollLeft is 63
PASS wheelEventsOccurred is 3
second gesture
wheel event 0+> [object HTMLDivElement]
wheel event 1+> [object HTMLDivElement]
received element removing keypress
PASS movingdiv.scrollTop is 0
PASS movingdiv.scrollLeft is 0
PASS wheelEventsOccurred is 2
scroll event 0+> [object HTMLDivElement]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
